remove rsi ticket search part from disputeApi.Web
