--- src/backend/TrafficCourtsApi/DisputeApi.Web/Auth/OAuthException.cs
@@ -1,27 +1,0 @@
-﻿using System;
-
-namespace DisputeApi.Web.Auth
-{
-    [Serializable]
-    public class OAuthException : System.Exception
-    {
-        public int StatusCode { get; private set; }
-
-        public string Response { get; private set; }
-
-        public System.Collections.Generic.IReadOnlyDictionary<string, System.Collections.Generic.IEnumerable<string>> Headers { get; private set; }
-
-        public OAuthException(string message, int statusCode, string response, System.Collections.Generic.IReadOnlyDictionary<string, System.Collections.Generic.IEnumerable<string>> headers, System.Exception innerException)
-            : base(message + "\n\nStatus: " + statusCode + "\nResponse: \n" + response.Substring(0, response.Length >= 512 ? 512 : response.Length), innerException)
-        {
-            StatusCode = statusCode;
-            Response = response;
-            Headers = headers;
-        }
-
-        public override string ToString()
-        {
-            return $"HTTP Response: \n\n{Response}\n\n{base.ToString()}";
-        }
-    }
-}
