--- src/backend/TrafficCourtsApi/DisputeApi.Web/Auth/OAuthHandler.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace DisputeApi.Web.Auth
-{
-    /// <summary>
-    /// The OAuthHandler intercept http request and add the OAuth token
-    /// </summary>
-    public class OAuthHandler : DelegatingHandler
-    {
-
-        private readonly ITokenService _tokenService;
-
-        public OAuthHandler(ITokenService tokenService)
-        {
-            _tokenService = tokenService;
-        }
-
-        protected override async Task<HttpResponseMessage> SendAsync(
-            HttpRequestMessage request,
-            CancellationToken cancellationToken)
-        {
-            var token = await _tokenService.GetTokenAsync(cancellationToken);
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", token.AccessToken);
-            return await base.SendAsync(request, cancellationToken);
-        }
-    }
-}
