--- src/backend/TrafficCourtsApi/DisputeApi.Web/Auth/OAuthOptions.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace DisputeApi.Web.Auth
-{
-    public class OAuthOptions
-    {
-        [Required]
-        public string OAuthUrl { get; set; }
-
-        [Required]
-        public string ResourceUrl { get; set; }
-
-        [Required]
-        public string ClientId { get; set; }
-
-        [Required]
-        public string Secret { get; set; }
-
-         [Range(1, 10)]
-        public int TokenTimeout { get; set; } = 1;
-    }
-}
