--- src/backend/TrafficCourtsApi/DisputeApi.Web/Auth/Token.cs
@@ -1,19 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace DisputeApi.Web.Auth
-{
-    public class Token
-    {
-        [JsonPropertyName("access_token")]
-        public string AccessToken { get; set; }
-
-        [JsonPropertyName("token_type")]
-        public string TokenType { get; set; }
-
-        [JsonPropertyName("expires_in")]
-        public int ExpiresIn { get; set; }
-
-        [JsonPropertyName("scope")]
-        public string Scope { get; set; }
-    }
-}
