--- src/backend/TrafficCourtsApi/DisputeApi.Web/Features/TicketLookup/ticket-EZ02000460.json
@@ -1,22 +1,0 @@
-{
-  "items": [
-    {
-      "selected_invoice": {
-        "$ref": "https://wsgw.test.jag.gov.bc.ca/paybc/vph/rest/PSSGVPHPAYBC/vph/invs/EZ020004601"
-      },
-      "open_invoices_for_site": null
-    },
-    {
-      "selected_invoice": {
-        "$ref": "https://wsgw.test.jag.gov.bc.ca/paybc/vph/rest/PSSGVPHPAYBC/vph/invs/EZ020004602"
-      },
-      "open_invoices_for_site": null
-    },
-    {
-      "selected_invoice": {
-        "$ref": "https://wsgw.test.jag.gov.bc.ca/paybc/vph/rest/PSSGVPHPAYBC/vph/invs/EZ020004603"
-      },
-      "open_invoices_for_site": null
-    }
-  ]
-}
